refactor(feature-flags): moves get util to avoid importing system context
